+++ app/application/common/use_case/auth_first_time_register.py
@@ -1,5 +1,5 @@
 from app.application.common.use_case.schemas.auth import FirstTimeRegisterIn, LoginOut
-from app.repositories.usuario_respository import UsuariosRepository
+from app.infrastructure.repositories.usuario_respository import UsuariosRepository
 
 class FirstTimeRegisterUseCase:
     def __init__(self, users: UsuariosRepository):
