+++ app/infrastructure/repositories/usuario_respository.py
@@ -41,24 +41,3 @@ class UsuariosRepository:
         )
         return self.db.execute(stmt).scalar_one_or_none()
     
-    def usuarios_del_conjunto(self, conjunto_id:int) -> List[dict]:
-            sql = text("""
-                SELECT u.id_usuario, u.tipo_usuario_id, u.administracion, a.id_apartamento
-                FROM usuario u
-                JOIN apartamento a ON a.usuario_id = u.id_usuario
-                WHERE a.conjunto_residencial_id = :cid
-            """)
-            with self.engine.begin() as con:
-                return [dict(r) for r in con.execute(sql, {"cid":conjunto_id}).mappings().all()]
-
-    def conteo_ganados(self, ids: List[int]) -> dict:
-        if not ids: return {}
-        sql = text("""
-                SELECT usuario_id, COUNT(*) AS c
-                FROM resultado_sorteo_detalle
-                WHERE usuario_id = ANY(:ids)
-                GROUP BY usuario_id
-            """)
-        with self.engine.begin() as con:
-                rows = con.execute(sql, {"ids": ids}).mappings().all()
-                return {r["usuario_id"]: r["c"] for r in rows}
