+++ app/interfaces/routers/setup_router.py
@@ -4,7 +4,7 @@ from app.infrastructure.db.session import get_db
 from app.application.common.use_case.schemas.setup_schemas import (
     SetupInitialOut, SetupInitialUpdateIn, SetupActionResult, TorresConfigIn
 )
-from app.repositories.setup_repository import SetupRepository
+from app.infrastructure.repositories.setup_repository import SetupRepository
 from app.application.common.use_case.setup_initial import (
     SetupInitialGetUseCase, SetupInitialUpdateUseCase, ConfigTorresUseCase
 )
